create routes in a new file to simpify server.js code

+++ backend/controllers/location.controller.js
@@ -35,7 +35,7 @@ export const updateLocation = async (req, res) => {
 	const location = req.body;
 
 	if (!mongoose.Types.ObjectId.isValid(id)) {
-		return res.status(400).json({ success: false, message: 'Invalid ID format' });
+		return res.status(404).json({ success: false, message: 'Invalid ID format' });
 	}
 
 	try {
@@ -50,11 +50,15 @@ export const updateLocation = async (req, res) => {
 export const deleteLocation = async (req, res) => {
 	const { id } = req.params;
 
+	if (!mongoose.Types.ObjectId.isValid(id)) {
+		return res.status(404).json({ success: false, message: 'Invalid ID format' });
+	}
+
 	try {
 		await Location.findByIdAndDelete(id);
 		res.status(200).json({ success: true, message: 'Location deleted' });
 	} catch (error) {
 		console.error('Error deleting location:', error.message);
-		res.status(404).json({ success: false, message: 'Location not found' });
+		res.status(500).json({ success: false, message: 'Server Error' });
 	}
 };
